+++ xlab/data/importer/iex/__init__.py
@@ -1,0 +1,1 @@
+from xlab.data.importer.iex.provider import IexDataImporter
