+++ xlab/data/importer/iex/importer_test.py
@@ -7,7 +7,7 @@ import requests
 from requests import exceptions as request_exceptions
 
 from xlab.data.proto import data_type_pb2
-from xlab.data.provider.iex import api, provider
+from xlab.data.importer.iex import api, importer
 from xlab.net.proto.testing import compare
 
 
@@ -21,13 +21,13 @@ def _make_response(conent: str,
     return response
 
 
-class IexDataProviderTest(absltest.TestCase):
+class IexDataImporterTest(absltest.TestCase):
 
     def setUp(self):
-        self.provider = provider.IexDataProvider(
+        self.importer = importer.IexDataImporter(
             api.IexApiHttpClient('fake_token'))
 
-    @patch.object(provider, 'datetime', Mock(wraps=datetime))
+    @patch.object(importer, 'datetime', Mock(wraps=datetime))
     @patch('requests.Session.get')
     def test_get_quotes_success(self, mock_get):
         api_response_data = """{
@@ -40,10 +40,10 @@ class IexDataProviderTest(absltest.TestCase):
         }"""
 
         mock_get.return_value = _make_response(api_response_data, 200)
-        provider.datetime.datetime.now.return_value = datetime.datetime(
+        importer.datetime.datetime.now.return_value = datetime.datetime(
             2020, 12, 31)
 
-        results = self.provider.get_data('SPY')
+        results = self.importer.get_data('SPY')
 
         _, kwargs = mock_get.call_args
         self.assertEqual(kwargs['params']['symbols'], 'SPY')
@@ -92,16 +92,16 @@ class IexDataProviderTest(absltest.TestCase):
     def test_get_quotes_default_end_date(self):
         pass
 
-    @patch.object(provider, 'datetime', Mock(wraps=datetime))
+    @patch.object(importer, 'datetime', Mock(wraps=datetime))
     @patch('requests.Session.get')
     def test_get_quotes_backend_failure(self, mock_get):
         mock_get.return_value = _make_response('', 400, 'Invalid argument!')
-        provider.datetime.datetime.now.return_value = datetime.datetime(
+        importer.datetime.datetime.now.return_value = datetime.datetime(
             2020, 12, 31)
 
         with self.assertRaisesRegex(request_exceptions.HTTPError,
                                     '400 Client Error: Invalid argument'):
-            self.provider.get_data('SPY')
+            self.importer.get_data('SPY')
 
     def test_get_quotes_invalid_dates(self):
         pass
