--- xlab/data/provider/iex/__init__.py
@@ -1,1 +1,0 @@
-from xlab.data.provider.iex.provider import IexDataProvider
